requirements.txt
py-cord>=2.6.0
openai-whisper
anthropic>=0.40.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
pydub
python-dotenv
PyNaCl
